+++ drift/test/integration_tests/database_test.dart
@@ -128,4 +128,27 @@ void main() {
 
     expect(await db.todosTable.all().get(), isEmpty);
   });
+
+  test('exclusively', () async {
+    final db = TodoDb(NativeDatabase.memory());
+    addTearDown(db.close);
+
+    bool query1Finished = false;
+
+    // Make a slow query and check that it is not interleaved with other queries/transactions
+    final fut1 = db.exclusively<int>(() async {
+      await Future<void>.delayed(const Duration(milliseconds: 100));
+      await db.customSelect('SELECT 1').get();
+      query1Finished = true;
+      return 7;
+    });
+
+    final fut2 = db.transaction<int>(() async {
+      expect(query1Finished, true);
+      return 5;
+    });
+
+    final results = await Future.wait<void>([fut1, fut2]);
+    expect(results, [7, 5]);
+  });
 }